+++ FinanceApp/src/Views/Components/ProfileParentView/TaskCreateView.swift
@@ -8,7 +8,7 @@ struct TaskCreateView: View {
     @ObservedObject private var parentViewModel = ParentViewModel()
     
     @State var taskDescription: String = ""
-    @State var value: Float = 0.0
+    @State var value: Int = 0
     @State var stringValue: String = ""
     @State var recurrent: Bool = true
     @State var selectedChild: ChildModel?
